Fix: 동의, 이메일폼 ViewModel 결합, 회원가입 ViewModel 수정
