+++ app/src/main/java/com/example/simya/src/ui/viewmodel/login/signup/SignupAgreeViewModel.kt
@@ -5,29 +5,6 @@ import androidx.lifecycle.ViewModel
 import androidx.lifecycle.viewModelScope
 import kotlinx.coroutines.launch
 
-class SignupAgreeViewModel : ViewModel() {
-
-    // 전체 약관 동의
-    var agreeAll = MutableLiveData<Boolean>()
-
-    // 개인정보 처리방침
-    var agreeInfo = MutableLiveData<Boolean>()
-
-    // 서비스 이용약관
-    var agreeService = MutableLiveData<Boolean>()
-
-
-    init{
-        agreeAll.value = false
-        agreeInfo.value = false
-        agreeService.value = false
-    }
-
-    fun isAgreeStatus(): Boolean{
-        return agreeInfo.value!! && agreeService.value!!
-    }
-    fun checkAgreeAllStatus() = viewModelScope.launch {
-        agreeService.value = agreeAll.value
-        agreeInfo.value = agreeAll.value
-    }
-}
+//class SignupAgreeViewModel : ViewModel() {
+//
+//}
