+++ app/src/main/java/com/example/simya/src/ui/viewmodel/login/signup/SignupViewModel.kt
@@ -2,8 +2,20 @@ package com.example.simya.src.ui.viewmodel.login.signup
 
 import androidx.lifecycle.MutableLiveData
 import androidx.lifecycle.ViewModel
+import androidx.lifecycle.viewModelScope
+import kotlinx.coroutines.launch
 
 class SignupViewModel: ViewModel() {
+
+
+    // 전체 약관 동의
+    var agreeAll = MutableLiveData<Boolean>()
+
+    // 개인정보 처리방침
+    var agreeInfo = MutableLiveData<Boolean>()
+
+    // 서비스 이용약관
+    var agreeService = MutableLiveData<Boolean>()
 
     val email = MutableLiveData<String>()
 
@@ -11,5 +23,20 @@ class SignupViewModel: ViewModel() {
 
     fun emailEmptyCheck():Boolean{
         return email.value.toString().isNotEmpty()
+    }
+
+
+    init{
+        agreeAll.value = false
+        agreeInfo.value = false
+        agreeService.value = false
+    }
+
+    fun isAgreeStatus(): Boolean{
+        return agreeInfo.value!! && agreeService.value!!
+    }
+    fun checkAgreeAllStatus() = viewModelScope.launch {
+        agreeService.value = agreeAll.value
+        agreeInfo.value = agreeAll.value
     }
 }
